requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
asgiref==3.11.0; python_version >= '3.9'
astroid==3.3.11; python_full_version >= '3.9.0'
autopep8==2.3.2; python_version >= '3.9'
certifi==2025.11.12; python_version >= '3.7'
charset-normalizer==3.4.4; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
contourpy==1.3.3; python_version >= '3.11'
cycler==0.12.1; python_version >= '3.8'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dill==0.4.0; python_version >= '3.8'
django==5.2.8; python_version >= '3.10'
djangorestframework==3.16.1; python_version >= '3.9'
filelock==3.20.0; python_version >= '3.10'
flake8==7.3.0; python_version >= '3.9'
fonttools==4.60.1; python_version >= '3.9'
fsspec==2025.10.0; python_version >= '3.9'
idna==3.11; python_version >= '3.8'
isort==6.1.0; python_full_version >= '3.9.0'
jinja2==3.1.6; python_version >= '3.7'
joblib==1.5.2; python_version >= '3.9'
kiwisolver==1.4.9; python_version >= '3.10'
lightning-utilities==0.15.2; python_version >= '3.9'
markupsafe==3.0.3; python_version >= '3.9'
matplotlib==3.10.7; python_version >= '3.10'
mccabe==0.7.0; python_version >= '3.6'
mpmath==1.3.0
networkx==3.5; python_version >= '3.11'
numpy==2.2.6; python_version >= '3.10'
opencv-python==4.12.0.88; python_version >= '3.6'
packaging==25.0; python_version >= '3.8'
pandas==2.3.3; python_version >= '3.9'
pillow==12.0.0; python_version >= '3.10'
platformdirs==4.5.0; python_version >= '3.10'
polars==1.35.2; python_version >= '3.9'
polars-runtime-32==1.35.2; python_version >= '3.9'
psutil==7.1.3; python_version >= '3.6'
pycodestyle==2.14.0; python_version >= '3.9'
pydantic==2.12.4; python_version >= '3.9'
pydantic-core==2.41.5; python_version >= '3.9'
pyflakes==3.4.0; python_version >= '3.9'
pylint==3.3.9; python_full_version >= '3.9.0'
pylint-django==2.6.1; python_version >= '3.9' and python_version < '4.0'
pylint-plugin-utils==0.9.0; python_version >= '3.9' and python_version < '4.0'
pyparsing==3.2.5; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-decouple==3.8
pytz==2025.2
pyyaml==6.0.3; python_version >= '3.8'
requests==2.32.5; python_version >= '3.9'
scikit-learn==1.7.2; python_version >= '3.10'
scipy==1.16.3; python_version >= '3.11'
setuptools==80.9.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.5.3; python_version >= '3.8'
supervision==0.27.0; python_version >= '3.9'
sympy==1.14.0; python_version >= '3.9'
threadpoolctl==3.6.0; python_version >= '3.9'
tomlkit==0.13.3; python_version >= '3.8'
torch==2.9.1; python_version >= '3.10'
torchmetrics==1.8.2; python_version >= '3.9'
torchvision==0.24.1; python_version >= '3.10'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.2; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
ultralytics==8.3.231; python_version >= '3.8'
ultralytics-thop==2.0.18; python_version >= '3.8'
urllib3==2.5.0; python_version >= '3.9'
